+++ Sources/TuistCache/ContentHashing/CacheGraphContentHasher.swift
@@ -60,7 +60,7 @@ public final class CacheGraphContentHasher: CacheGraphContentHashing {
             additionalStrings: [
                 cacheProfileContentHasher.hash(cacheProfile: cacheProfile),
                 cacheOutputType.description,
-                xcodeBuildController.version().toBlocking().single().value,
+                xcodeBuildController.version().toBlocking().single(),
                 Constants.version,
             ]
         )